requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
supabase
kafka-python==2.0.2
pydantic>=2.12,<3
fastapi>=0.120.1
uvicorn==0.22.0
cloudinary==1.30.0
gradio==3.50.2
